revert reorg
